# Click-Reserva/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ClickReserva API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms:
    get:
      operationId: getRooms
      tags: [rooms]
      summary: List all rooms
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a new room (coordinator only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomBody"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{roomId}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get a single room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    put:
      operationId: updateRoom
      tags: [rooms]
      summary: Update a room (coordinator only)
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomBody"
      responses:
        "200":
          description: Room updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete a room (coordinator only)
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Room:
      type: object
      properties:
        id:
          type: integer
        number:
          type: string
        name:
          type: string
        capacity:
          type: integer
        computers:
          type: integer
        isActive:
          type: boolean
        description:
          type: string
          nullable: true
      required:
        - id
        - number
        - name
        - capacity
        - computers
        - isActive
    CreateRoomBody:
      type: object
      properties:
        number:
          type: string
        name:
          type: string
        capacity:
          type: integer
        computers:
          type: integer
        isActive:
          type: boolean
        description:
          type: string
          nullable: true
      required:
        - number
        - name
        - capacity
        - computers
...